Add support for the -exclude flag
Excludes are specified in the configuration file, and the expected file
for errcheck is created on disk as needed. The upgrader does NOT attempt
to covert configurations that used the -exclude flag.

This also changes the behavior of the ignore flag to join each argument
with a comma, since errcheck only accepts a single argument for this
flag. This shouldn't break existing users, as any project specifying
multiple patterns was already broken.